core/standards/css3/image-fit/js/ua-stylesheet/embed/004.html

<!doctype html>
<title>embed default style, video</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<embed src=../../support/240x160.ogv>
<script>
test(function() {
  var e = document.body.firstChild;
  assert_equals(getComputedStyle(e, '').OObjectFit, 'auto');
  assert_equals(getComputedStyle(e, '').OObjectPosition, '50% 50%');
}, window, 'load', 2000);
</script>


